Add extension for better search
The original search shows found tasks in a new TaskList, this means
that the items are not displayed with their corresponding index. This
makes it difficult for the user to mark/delete the task after found.

The new implementation shows the tasks found along with its
corresponding index by creating a new record TaskMatch, which include
the task and its index in the original TaskList.

This approach has some downside in that passing around a
List<TaskMatch> object around is not so OOP, but for a single method,
this implementation is sufficient.